refactor: 컨테이너 컴포넌트 템플릿으로 이동
